% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_model.R
\name{analyze_model}
\alias{analyze_model}
\title{A function that analyzes the model() object created or loaded in the main menu.}
\usage{
analyze_model(modeltype, rngseed, nreps, plotscale, input, model)
}
\arguments{
\item{modeltype}{A string indicating the type of model. Accepted values are "ode", "stochastic", and "discrete".}

\item{rngseed}{A random number seed for the simulation.}

\item{nreps}{Number of times to run the simulation.}

\item{plotscale}{Log or linear scale for the x-axis, y-axis, or both.}

\item{input}{The Shiny input list generated in the main menu app.}

\item{model}{The Shiny model() object generated in the main menu app.}
}
\value{
A list named "result" with the simulated dataframe and associated metadata.
}
\description{
This function takes the model specified in the main menu and
runs the simulation determined by the model parameters.
}
\details{
This function is called by the Shiny server to produce the Shiny input UI elements.
}
\author{
Spencer D. Hall
}
