% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_from_desolve.R
\name{convert_from_desolve}
\alias{convert_from_desolve}
\title{Create model from desolve function}
\usage{
convert_from_desolve(desolvefunction)
}
\arguments{
\item{desolvefunction}{name of file that should be converted}
}
\value{
A list structure containing the model
}
\description{
This function performs the inverse action of convert_to_desolve()
If a model is turned into a desolve function with convert_to_desolve and then run through this function
the result should be identical to the starting model structure
comparison can be done with the setdiff() function
}
\details{
This function takes as input a desolve function code and turns it into a model list structure
}
\author{
Andreas Handel
}
