% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_desolve.R
\name{convert_to_desolve}
\alias{convert_to_desolve}
\title{Create a desolve function}
\usage{
convert_to_desolve(model, location)
}
\arguments{
\item{model}{model structure, either as list or Rdata file name}

\item{location}{a path/folder to save the model to. Default is current directory}
}
\value{
The function does not return anything
Instead, it writes an R file into the working directory
this R file contains a desolve implementation of the model
the name of the file is model$title_desolve.R
}
\description{
USER CAN ADD MORE DETAILS HERE
}
\details{
This function takes as input a model and writes an R desolve function
}
\author{
Andreas Handel
}
