% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_interaction_flows.R
\name{find_interaction_flows}
\alias{find_interaction_flows}
\title{Identifies flows with interactions among state variables}
\usage{
find_interaction_flows(mbmodel)
}
\arguments{
\item{mbmodel}{modelbuilder model structure, either as list object
or file name}
}
\value{
The function returns a data frame identifying the flows (and
    their associated vars) with state variable interactions.
}
\description{
The model needs to adhere to the structure specified by
    the modelbuilder package. Models built using the modelbuilder package
    automatically have the right structure. A user can also build a
    model list structure themselves following the modelbuilder
    specifications. If the user provides a file name, this file needs to
    be an RDS file and contain a valid modelbuilder model structure.
}
\details{
This function takes as input a modelbuilder model and
identifies the flows that have interactions among state variables.
This is an internal function used by the generate_stratified_model
function.
}
\author{
Andrew Tredennick and Andreas Handel
}
