% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_modelsettings.R
\name{find_modelsettings}
\alias{find_modelsettings}
\title{A helper function that takes a modelbuilder UI and extracts the model settings}
\usage{
find_modelsettings(input, mbmodel)
}
\arguments{
\item{input}{Shiny UI containing model settings}

\item{mbmodel}{a modelbuilder model structure}
}
\value{
modelsettings a list containg variables for model settings
}
\description{
This function takes a modelbuilder input structure as generated by
other functions in that package and extracts UI inputs and saves in a model structure
This is a helper function.
}
\details{
Returned list has vectors var par an time for the model setting.
Also returned are UI inputs modeltype, rngseed, nreps, plotscae
}
\author{
Andreas Handel
}
