% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_buildUI.R
\name{generate_buildUI}
\alias{generate_buildUI}
\title{A helper function that takes a model and generates the shiny UI elements for the build tab}
\usage{
generate_buildUI(mbmodel, output, values)
}
\arguments{
\item{mbmodel}{a modelbuilder model structure}

\item{output}{shiny output structure}

\item{values}{global variable to track number of vars/pars/flows}
}
\value{
No direct return. output structure is modified to contain text for display in a Shiny UI
}
\description{
This function generates empty inputs or inputs for a supplied model.
This is a helper function called by the shiny app.
}
\details{
This function is called by the Shiny server to produce the Shiny input UI elements for the build tab.
}
\author{
Andreas Handel
}
