% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diagram.R
\name{generate_diagram}
\alias{generate_diagram}
\title{Create a diagram for a model}
\usage{
generate_diagram(model)
}
\arguments{
\item{model}{model structure, either as list object or Rdata file name}
}
\value{
The function returns the diagram stored in a variable
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package
models built using the modelbuilder package automatically have the right structure
a user can also build a model list structure themselves following the specifications
if the user provides an Rdata file name, this file needs to contain an object called 'model'
and contain a valid modelbuilder model structure
}
\details{
This function takes as input a modelbuilder model
and creates a diagram using ggplot
}
\author{
Andreas Handel
}
