% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_discrete.R
\name{generate_discrete}
\alias{generate_discrete}
\title{Create a discrete time simulation model}
\usage{
generate_discrete(mbmodel, location = NULL, filename = NULL)
}
\arguments{
\item{mbmodel}{modelbuilder model structure, either as list object or file name}

\item{location}{a path to save the simulation code to. If NULL, defaults to current directory.}

\item{filename}{a filename to save the simulation code to. If NULL, a default is generated (recommended).}
}
\value{
The function does not return anything
Instead, it writes an R file into the specified directory
the default name of the file is simulate_model$title_ode.R
if the user specifies a file name, it will be that name
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package.
Models built using the modelbuilder package automatically have the right structure.
A user can also build a model list structure themselves following the modelbuilder specifications.
If the user provides a file name, this file needs to be an Rds file
and contain a valid modelbuilder model structure.
}
\details{
This function takes as input a modelbuilder model and writes code
for a discrete time deterministic simulator function
}
\author{
Andreas Handel
}
