% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_flowchart.R
\name{generate_flowchart}
\alias{generate_flowchart}
\title{This function takes as input a modelbuilder model
    and creates a diagram using DiagrammeR.}
\usage{
generate_flowchart(mbmodel, code_path = NULL)
}
\arguments{
\item{mbmodel}{model structure, either as list object or Rds file name}

\item{code_path}{path where ggplot2 code will be written}
}
\value{
The function returns the diagram stored in a variable
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package
models built using the modelbuilder package automatically have the right structure
a user can also build a model list structure themselves following the specifications
if the user provides an Rds file name, this file needs to contain an object called 'mbmodel'
and contain a valid modelbuilder model structure
}
\author{
Andreas Handel and Andrew Tredennick
}
