% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_flowchart_ggplot.R
\name{generate_flowchart_ggplot}
\alias{generate_flowchart_ggplot}
\title{This function takes as input a modelbuilder model and creates a diagram using ggplot}
\usage{
generate_flowchart_ggplot(model)
}
\arguments{
\item{model}{model structure, either as list object or Rdata file name}
}
\value{
The function returns the diagram stored in a variable
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package
models built using the modelbuilder package automatically have the right structure
a user can also build a model list structure themselves following the specifications
if the user provides an Rdata file name, this file needs to contain an object called 'model'
and contain a valid modelbuilder model structure
}
\author{
Andreas Handel
}
