% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_inputs.R
\name{generate_function_inputs}
\alias{generate_function_inputs}
\title{Create function input/arguments part for a simulation model}
\usage{
generate_function_inputs(mbmodel, modeltype)
}
\arguments{
\item{mbmodel}{modelbuilder model structure}

\item{modeltype}{Indicates type of model code that is generated. One of ode/stochastic/discrete}
}
\value{
text that to be placed on top of function code generated by the calling function
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package.
Models built using the modelbuilder package automatically have the right structure.
A user can also build a model list structure themselves following the modelbuilder specifications.
If the user provides a file name, this file needs to be an Rds file
and contain a valid modelbuilder model structure.
}
\details{
This function takes as input a modelbuilder model and writes code
that produces text needed to create the main function call and the var/par/time vectors
}
