% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{generate_model}
\alias{generate_model}
\title{A helper function that takes a model as a list structure and generates a tar file with code}
\usage{
generate_model(model, location = NULL)
}
\arguments{
\item{model}{a modelbuilder model structure}

\item{location}{a path/folder to save the model to. Default is modelexamples subfolder of package}
}
\value{
no return, the function writes a tar file with the model code
}
\description{
This function generates a tar file that includes the model as Rdata file
and code for the model implemented as discrete-time, stochastic/adaptivetau and ODE/desolve functions
}
\author{
Andreas Handel
}
