% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model_file.R
\name{generate_model_file}
\alias{generate_model_file}
\title{Create code that encodes a modelbuilder model}
\usage{
generate_model_file(mbmodel, location = NULL, filename = NULL)
}
\arguments{
\item{mbmodel}{modelbuilder model structure, either as list object or file name (with complete path)}

\item{location}{a path to save the code to. If NULL, defaults to current directory.}

\item{filename}{a filename to save the code to. If NULL, a default is generated (recommended).}
}
\value{
The function does not return anything
Instead, it writes an R file into the specified directory
the default name of the file is model$title_file.R
if the user specifies a file name, it will be that name
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package.
Models built using the modelbuilder package automatically have the right structure.
A user can also build a model list structure themselves following the modelbuilder specifications.
The file name must contain a valid modelbuilder Rds file.
}
\details{
This function takes as input an Rds file that has a modelbuilder model stored as list
it translates the list into generative code for easy, non-GUI edits to a model
}
\author{
Andreas Handel
}
