% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stochastic.R
\name{generate_stochastic}
\alias{generate_stochastic}
\title{Create a stochastic simulation model}
\usage{
generate_stochastic(model, location)
}
\arguments{
\item{model}{model structure, either as list object or Rdata file name}

\item{location}{a path/folder to save the simulation code to.}
}
\value{
The function does not return anything.
Instead, it writes an R file into the specified directory.
This R file contains a rxode implementation of the model.
the name of the file is simulate_model$title_stochastic.R
}
\description{
The model needs to adhere to the structure specified by the modelbuilder package
models built using the modelbuilder package automatically have the right structure
a user can also build a model list structure themselves following the specifications
if the user provides an Rdata file name, this file needs to contain an object called 'model'
and contain a valid modelbuilder model structure
}
\details{
This function takes as input a modelbuilder model and writes code
for a stochastic simulator function using the adpativetau package
}
\author{
Ishaan Dave, Andreas Handel
}
