% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stratified_model.R
\name{generate_stratified_model}
\alias{generate_stratified_model}
\title{Stratify an mb model by groups}
\usage{
generate_stratified_model(mbmodel, stratum_list, par_stratify_list)
}
\arguments{
\item{mbmodel}{modelbuilder model structure, either as list object
or file name}

\item{stratum_list}{A list of lists defining the stratification structure
to be applied to the model. At a minimum, the list must contain one
sublist with the following structure: \code{stratumname} a character
string giving the name of the stratum (e.g., "age"); \code{names} a
vector of character strings defining the names for each group within
the stratum (e.g., c("child", "adult")); \code{labels} a vector of
character strings defining the labels that correspond to each group
within the stratum, which are applied to the model codes (e.g.,
c("c", "a")); \code{comment} a character string of any comments or
notes for the stratum. The \code{labels} are appended to model state
variables (e.g., S becomes S_c) and parameters (e.g., b becomes b_c).}

\item{par_stratify_list}{A list that defines the stratification of
parameters in reference to state variables. This list can be auto
generated by calling the modelbuilder \code{generate_stratifier_list}.}
}
\value{
The function returns a modelbuilder model object (a list).
}
\description{
The model needs to adhere to the structure specified by
    the modelbuilder package. Models built using the modelbuilder package
    automatically have the right structure. A user can also build a
    model list structure themselves following the modelbuilder
    specifications. If the user provides a file name, this file needs to
    be an RDS file and contain a valid modelbuilder model structure.
}
\details{
This function takes as input a modelbuilder model and
expands it according to specified stratifications (e.g.
age groups).
}
\examples{
\dontrun{
mbmodel <- readRDS("auxiliary/modelfiles/Coronavirus_vaccine_model_v2.Rds")
strata_list <- list(
 list(
   stratumname = "age",
   names = c("children", "adults", "elderly"),
   labels = c("c", "a", "e"),
   comment = "This defines the age structure."
 ),
 list(
   stratumname = "risk",
   names = c("high risk", "low risk"),
   labels = c("h", "l"),
   comment = "This defines the risk structure."
 )
)
}
}
\author{
Andrew Tredennick and Andreas Handel
}
