% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{generate_tables}
\alias{generate_tables}
\title{Generate a table of state variables, parameters, and initial values}
\usage{
generate_tables(mbmodel, location = NULL, filename = NULL)
}
\arguments{
\item{mbmodel}{modelbuilder model structure, either as list object
or file name}

\item{location}{a path to save the simulation code to. If NULL,
defaults to current directory.}

\item{filename}{a filename to save the simulation code to. If NULL,
a default is generated (recommended).}
}
\value{
The function does not return anything
    Instead, it writes a CSV file into the specified directory;
    the default nams of the files are parameter_table_model$title.csv and
    parameter_table_model$title.csv; if the user specifies a file name,
    it will be that name
}
\description{
The model needs to adhere to the structure specified by
    the modelbuilder package. Models built using the modelbuilder package
    automatically have the right structure. A user can also build a
    model list structure themselves following the modelbuilder
    specifications. If the user provides a file name, this file needs to
    be an RDS file and contain a valid modelbuilder model structure.
}
\details{
This function takes as input a modelbuilder model and
exports a CSV table of parameter/variable names, definitions,
and initial values.
}
\author{
Andrew Tredennick and Andreas Handel
}
