% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_math.R
\name{get_math}
\alias{get_math}
\title{Extract the math symbols from a flow}
\usage{
get_math(flow, flowsymbols)
}
\arguments{
\item{flow}{A modelbuilder flow, which is a character string.}

\item{flowsymbols}{A character vector of the flow variables and parameters,
as returned from \code{get_vars_pars}.}
}
\value{
A character vector of the math symbols, in order.
}
\description{
The flow must be a character string of typical modelbuilder
    model structure. The function also concatenates parentheses and
    the notation before and after them as necessary.
}
\details{
This function takes as input a modelbuilder flow and
extracts the math symbols/notation.
}
\author{
Andrew Tredennick and Andreas Handel
}
