% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch a shiny app and pass an object to it}
\usage{
launch_app(modelfile)
}
\arguments{
\item{modelfile}{Name of a model/Rdata file}
}
\value{
Nothing
}
\description{
USER CAN ADD MORE DETAILS HERE
}
\details{
This function takes as input an object and makes it available inside the shiny app
}
\author{
Andreas Handel
}
