% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{A function that loads a model saved as Rdata file}
\usage{
load_model(currentmodel)
}
\arguments{
\item{currentmodel}{A file name and location for the model to be loaded.}
}
\value{
The model object stored in the Rdata file as a reactive
}
\description{
This function takes the name of an Rdata file that contains a modelbuilder model
and loads it as a reactive model
}
\details{
This function is a helper function
}
\author{
Spencer D. Hall, Andreas Handel
}
