% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_model_flow.R
\name{remove_model_flow}
\alias{remove_model_flow}
\title{A helper function that removes a the shiny UI elements for a flow of the build tab}
\usage{
remove_model_flow(values, input, output)
}
\arguments{
\item{values}{a shiny variable keeping track of UI elements}

\item{input}{shiny input structure}

\item{output}{shiny output structure}
}
\value{
No direct return. output structure is modified to contain text for display in a Shiny UI
}
\description{
This function removes inputs for the flows of a specified variable.
This is a helper function called by the shiny app.
}
\details{
This function is called by the Shiny server to produce the Shiny input UI elements for the build tab.
}
\author{
Andreas Handel
}
