% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_model_variableModal.R
\name{remove_model_variableModal}
\alias{remove_model_variableModal}
\title{produces remove variable confirmation modal}
\usage{
remove_model_variableModal(varName)
}
\arguments{
\item{varName}{Character value of the name of the variable in question.}
}
\description{
This modal asks users for confirmation that a given variable should be removed.
}
\author{
Joe Zientek - WEST, Inc
}
