% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_model_desolve.R
\name{SIR_model_desolve}
\alias{SIR_model_desolve}
\title{SIR model}
\usage{
SIR_model_desolve(vars = c(S = 1000, I = 1, R = 0), pars = c(b = 0.002,
  g = 1), times = c(tstart = 0, tfinal = 100, dt = 0.1))
}
\arguments{
\item{vars}{vector of starting conditions for these model variables: 
S : starting value for Susceptible
I : starting value for Infected
R : starting value for Recovered}

\item{pars}{vector of values for these model parameters: 
b : infection rate
g : recovery rate}

\item{times}{vector of values for these model times: 
tstart : Start time of simulation
tfinal : Final time of simulation
dt : Time step}
}
\value{
The function returns the output as a list. 
The time-series from the simulation is returned as a dataframe saved as list element ts. 
The ts dataframe has one column per compartment/variable. The first column is time.
}
\description{
USER CAN ADD MORE DETAILS HERE
}
\details{
A basic SIR model with 3 compartments and infection and recovery processes

USER CAN ADD MORE DETAILS HERE
}
\section{Warning}{
 This function does not perform any error checking. So if you try to do something nonsensical (e.g. have negative values for parameters), the code will likely abort with an error message
}

\section{Model Author}{
 Andreas Handel
}

\section{Model creation date}{
 2018-09-04
}

\section{Function Author}{
 generated by convert_to_desolve function
}

\section{Function creation date}{
 2018-09-07
}

\examples{
 
# To run the simulation with default parameters just call the function:  
result <- SIR_model_desolve() 
}
