% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vector_transmission_model_desolve.R
\name{Vector_transmission_model_desolve}
\alias{Vector_transmission_model_desolve}
\title{Vector transmission model}
\usage{
Vector_transmission_model_desolve(vars = c(Sh = 1000, Ih = 1, Rh = 0, Sv
  = 1000, Iv = 1), pars = c(b1 = 0.002, b2 = 0.002, g = 1, w = 0.1, b =
  100, n = 0.1), time = c(tstart = 0, tfinal = 100, dt = 0.1))
}
\arguments{
\item{Sh}{starting value for Susceptible Humans}

\item{Ih}{starting value for Infected Humans}

\item{Rh}{starting value for Recovered Humans}

\item{Sv}{starting value for Susceptible Vectors}

\item{Iv}{starting value for Infected Vctors}

\item{b1}{infection rate of hosts}

\item{b2}{infection rate of vectors}

\item{g}{recovery rate of hosts}

\item{w}{wanning immunity rate}

\item{b}{vector birth rate}

\item{n}{vector death rate}

\item{tstart}{Start time of simulation}

\item{tfinal}{Final time of simulation}

\item{dt}{Time step}
}
\value{
The function returns the output as a list. 
The time-series from the simulation is returned as a dataframe saved as list element ts. 
The ts dataframe has one column per compartment/variable. The first column is time.
}
\description{
USER CAN ADD MORE DETAILS HERE
}
\details{
A basic model with several compartments to model vector-borne transmission

USER CAN ADD MORE DETAILS HERE
}
\examples{
 
# To run the simulation with default parameters just call the function:  
result <- Vector_transmission_model_desolve() 
}
