% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{generate_model}
\alias{generate_model}
\title{A helper function that reads the shiny model builder UI inputs and saves information in a model object}
\usage{
generate_model(input, values)
}
\arguments{
\item{input}{Shiny UI inputs to be read}

\item{values}{Shiny UI variable that contains needed counters}
}
\value{
A modelbuilder model structure.
}
\description{
This function takes as input the shiny user inputs for the modelbuilder build functionality
and writes them inot a model object
}
\details{
This function is a helper function
}
\author{
Spencer D. Hall, Andreas Handel
}
