% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_model_var.R
\name{remove_model_var}
\alias{remove_model_var}
\title{A helper function that removes a variable to the shiny UI elements of the build tab}
\usage{
remove_model_var(mbmodel, values, input, output)
}
\arguments{
\item{mbmodel}{a modelbuilder model structure}

\item{values}{a shiny variable keeping track of UI elements}

\item{input}{shiny input structure}

\item{output}{shiny output structure}
}
\value{
No direct return. output structure is modified to contain text for display in a Shiny UI
}
\description{
This function removes inputs for the last variable.
This is a helper function called by the shiny app.
}
\details{
This function is called by the Shiny server to produce the Shiny input UI elements for the build tab.
}
\author{
Andreas Handel
}
