% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vector_transmission_model_desolve.R
\name{Vector_transmission_model_desolve}
\alias{Vector_transmission_model_desolve}
\title{Vector transmission model}
\usage{
Vector_transmission_model_desolve(vars = c(Sh = 1000, Ih = 1, Rh = 0, Sv
  = 1000, Iv = 1), pars = c(b1 = 0.002, b2 = 0.002, g = 1, w = 0.1, b =
  100, n = 0.1), times = c(tstart = 0, tfinal = 100, dt = 0.1))
}
\arguments{
\item{vars}{vector of starting conditions for model variables: 
\itemize{ 
\item Sh : starting value for Susceptible hosts
\item Ih : starting value for Infected hosts
\item Rh : starting value for Recovered hosts
\item Sv : starting value for Susceptible Vectors
\item Iv : starting value for Infected Vectors
}}

\item{pars}{vector of values for model parameters: 
\itemize{ 
\item b1 : infection rate of hosts
\item b2 : infection rate of vectors
\item g : recovery rate of hosts
\item w : wanning immunity rate
\item b : vector birth rate
\item n : vector death rate
}}

\item{times}{vector of values for model times: 
\itemize{ 
\item tstart : Start time of simulation
\item tfinal : Final time of simulation
\item dt : Time step
}}
}
\value{
The function returns the output as a list. 
The time-series from the simulation is returned as a dataframe saved as list element \code{ts}. 
The \code{ts} dataframe has one column per compartment/variable. The first column is time.
}
\description{
A basic model with several compartments to model vector-borne transmission
}
\details{
The model tracks the dynamics of susceptible, infected, and recovered hosts, and susceptible and infected vectors. Infection, recovery, and waning immunity processes are implemented for hosts. Births and deaths and infection processes are implemented for vectors.
}
\section{Warning}{
 This function does not perform any error checking. So if you try to do something nonsensical (e.g. have negative values for parameters), the code will likely abort with an error message
}

\section{Model Author}{
 Andreas Handel
}

\section{Model creation date}{
 2018-09-07
}

\section{Code Author}{
 generated by the \code{convert_to_desolve} function
}

\section{Code creation date}{
 2018-09-07
}

\examples{
 
# To run the simulation with default parameters:  
result <- Vector_transmission_model_desolve() 
}
